% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trm_colors.R
\name{trm_colors}
\alias{trm_colors}
\title{Farver i TRM}
\usage{
trm_colors(farver, lys = 0)
}
\arguments{
\item{farver}{Vælg mellem c("blaa", "cyan", "orange", "groen", "roed")}

\item{lys}{Lyshedsgrad: 0 (fuld), 1, 2 eller 3}
}
\description{
Farver i TRM
}
