% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trm_colors.R
\name{trm_colors}
\alias{trm_colors}
\title{Farver i TRM}
\usage{
trm_colors(farver)
}
\arguments{
\item{farver}{Choose between these colors c("blaa", "orange","graa", "gul","groen", "lyseblaa")}
}
\description{
Farver i TRM
}
