% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trm_colors.R
\name{trm_colors}
\alias{trm_colors}
\title{Farver i TRM}
\usage{
trm_colors(farver)
}
\arguments{
\item{farver}{Choose between these colors c("blå", "orange","grå", "gul","grøn", "lyseblå")}
}
\description{
Farver i TRM
}
